# openapi.yaml
openapi: "3.0.3"
info:
  title: WebPeel API
  description: |
    Fast web fetcher for AI agents. Convert any website to LLM-ready markdown, 
    search the web, crawl sites, extract structured data, and more.
    
    **Base URL:** `https://api.webpeel.dev`
    
    **Authentication:** Optional. Pass API key via `Authorization: Bearer wp_...` header.
    Free tier available without authentication (125 requests/week).
    
    **Firecrawl Compatible:** WebPeel supports Firecrawl-compatible endpoints at the same paths.
    Switch from Firecrawl by changing your base URL.
  version: "0.8.1"
  contact:
    name: WebPeel Support
    email: support@webpeel.dev
    url: https://webpeel.dev
  license:
    name: AGPL-3.0-or-later
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://api.webpeel.dev
    description: Production API
  - url: http://localhost:3000
    description: Local / Self-hosted
security:
  - bearerAuth: []
  - {}
paths:
...
  /v1/batch/scrape:
    post:
      operationId: batchScrape
      summary: Batch scrape multiple URLs
      tags: [Batch]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [urls]
              properties:
                urls:
                  type: array
                  items:
                    type: string
                  description: URLs to scrape
                options:
                  type: object
                  properties:
                    format:
                      type: string
                    render:
                      type: boolean
                    stealth:
                      type: boolean
      responses:
        "200":
          description: Batch job started
  /v1/batch/scrape/{id}:
    get:
      operationId: getBatchStatus
      summary: Get batch job status
      tags: [Batch]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Batch status and results
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: API key (wp_...) or Firecrawl key (fc-...)
...